chore: merge latest changes (#32)
* feat!: add Prisma integration (#22)

Adds a new Interface object to a `db.ts` and support for a Prisma database(through `postgresql`). There may be bugs and/or breaking changes.
(Also fixes a typo in the docs.)

* chore: run pre-build checks (#25)

* chore: merge new changes

* feat: basic prisma integration

* docs: fix typo


Rename ‘architechture’ to ‘architecture’.

* feat: added Interface object to represent prisma database

* chore: fix my bad code

* fix: return actual db entry instead of `true` (#26)

Returns the correct db entry for the channel instead of just `true`.

* docs: add documentation for adding database

A database is required to run Prisma, informs how to add your own `postgresql` server to the program.

* feat: dbres type for queries (#27)

Adds a `dbres` type which can be used to process the results of queries.

* Update db.ts

* fix: fix up addAutoThread 

Also adds spaces to comments.

* fix: proper types for db queries

* fix: face actually returns boolean (#28)

Added some fixes to make the databse interface work.

* feat: added the framework for the pin system, and create the base for
the event file. (note this does not include any actual code, just todos)

* chore: switch to eslint

* feat: search_repo prototype feature

* ci: add linting stage, run on push and PRs

* ci: hotfix ci by *not* using clean-install

Co-authored-by: BlueFalconHD <72631767+BlueFalconHD@users.noreply.github.com>
Co-authored-by: Andreas Grafen <andreas@grafen.info>
Co-authored-by: winston <hey@winston.sh>